feat(create-plugin): use react.lazy to code split app templates
