+++ packages/create-plugin/templates/app/src/components/App/App.tsx
@@ -2,19 +2,22 @@ import React from 'react';
 import { Route, Routes } from 'react-router-dom';
 import { AppRootProps } from '@grafana/data';
 import { ROUTES } from '../../constants';
-import { PageFour, PageOne, PageThree, PageTwo } from '../../pages';
+const PageOne = React.lazy(() => import('../../pages/PageOne'));
+const PageTwo = React.lazy(() => import('../../pages/PageTwo'));
+const PageThree = React.lazy(() => import('../../pages/PageThree'));
+const PageFour = React.lazy(() => import('../../pages/PageFour'));
 
 export function App(props: AppRootProps) {
   return (
-      <Routes>
-        <Route path={ROUTES.Two} element={<PageTwo />} />
-        <Route path={`${ROUTES.Three}/:id?`} element={<PageThree />} />
+    <Routes>
+      <Route path={ROUTES.Two} element={<PageTwo />} />
+      <Route path={`${ROUTES.Three}/:id?`} element={<PageThree />} />
 
-        {/* Full-width page (this page will have no side navigation) */}
-        <Route path={ROUTES.Four} element={<PageFour />} />
+      {/* Full-width page (this page will have no side navigation) */}
+      <Route path={ROUTES.Four} element={<PageFour />} />
 
-        {/* Default page */}
-        <Route path="*" element={<PageOne />} />
-      </Routes>
+      {/* Default page */}
+      <Route path="*" element={<PageOne />} />
+    </Routes>
   );
 }
